+++ src/converters/piral-cycle/src/converter.ts
@@ -8,7 +8,7 @@ export function createConverter() {
   return <TProps extends BaseComponentProps, M extends MatchingMain<PiralDomDrivers<TProps>, M>>(
     main: M,
   ): ForeignComponent<TProps> => {
-    const props$ = xs.create<TProps>();
+    let props$ = xs.create<TProps>();
     let dispose = () => {};
 
     return {
@@ -33,6 +33,7 @@ export function createConverter() {
       unmount() {
         props$.shamefullySendComplete();
         dispose();
+        props$ = xs.create<TProps>();
       },
     };
   };